+++ src/adapters/missing-adapter.ts
@@ -4,7 +4,7 @@ import type { Registry } from '../lib/library-types.ts';
 export interface IMissingAdapter {
   parseLibraryFolders(): Promise<Record<string, any>>;
   getRegistry(): Promise<Registry>;
-  computeDependencies(library: string, mode: string, version: null, folder?: string): Promise<Record<string, any>>;
+  computeDependencies(library: string, mode: 'view' | 'edit', version: string | null, folder?: string): Promise<Record<string, any>>;
 }
 
 export class MissingAdapter implements IMissingAdapter {
@@ -16,7 +16,7 @@ export class MissingAdapter implements IMissingAdapter {
     return logic.getRegistry();
   }
 
-  computeDependencies(library: string, mode: 'view' | 'edit', version: null, folder?: string): Promise<Record<string, any>> {
+  computeDependencies(library: string, mode: 'view' | 'edit', version: string | null, folder?: string): Promise<Record<string, any>> {
     return logic.computeDependencies(library, mode, version, folder);
   }
 }